add:create confirm

+++ src/backend/books/models.py
@@ -1,5 +1,6 @@
 from sqlalchemy import Boolean, Column, ForeignKey, Integer, String
 from sqlalchemy.orm import relationship
+from sqlalchemy.sql.sqltypes import Float
 
 from db import BasicModel
 
@@ -24,7 +25,7 @@ class Chapter(BasicModel):
   author = Column(String, index=True, comment='著者')
   word_count = Column(Integer, comment='文字数')
   e_pub = Column(String, nullable=False, index=True, comment='ePubデータ保存先')
-  matrix_row = Column(Integer, nullable=False, index=True, comment='内容傾向マトリックスの行番号')
+  matrix_row = Column(Float, nullable=False, index=True, comment='内容傾向マトリックスの行番号')
   book_id = Column(Integer, ForeignKey("books.id"))
 
   book = relationship("Book", back_populates="chapters")
